# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: screener
    description: Stock screener operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /screener/run:
    post:
      operationId: runScreener
      tags: [screener]
      summary: Run the stock screener
      description: Screens a list of tickers against P/E and profit margin criteria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenerRequest"
      responses:
        "200":
          description: Screener results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenerResponse"
  /screener/quote:
    post:
      operationId: getStockQuotes
      tags: [screener]
      summary: Get stock quotes for a list of tickers
      description: Returns current price, P/E, and margin for each ticker without filtering
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequest"
      responses:
        "200":
          description: Stock quotes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
  /screener/top-marketcap:
    post:
      operationId: getTopByMarketCap
      tags: [screener]
      summary: Get top US stocks by market cap
      description: Returns up to 1000 US equity tickers sorted by market cap descending using Yahoo Finance screener
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Top stocks by market cap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopMarketCapResponse"
  /screener/search-futures:
    post:
      operationId: searchFutures
      tags: [screener]
      summary: Search for futures/commodity ticker symbols by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TickerSearchRequest"
      responses:
        "200":
          description: Futures search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TickerSearchResponse"
  /screener/benchmarks:
    get:
      operationId: getBenchmarks
      tags: [screener]
      summary: Get major U.S. equity benchmark quotes
      description: Returns current price and day change for S&P 500, DJIA, Nasdaq, and Russell 2000
      responses:
        "200":
          description: Benchmark quotes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarksResponse"
  /screener/rates:
    get:
      operationId: getRates
      tags: [screener]
      summary: Get current fixed income and rates quotes
      description: Returns current quotes for US Treasury yields and MBS proxy
      responses:
        "200":
          description: Current rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatesResponse"
  /screener/search:
    post:
      operationId: searchTickers
      tags: [screener]
      summary: Search for ticker symbols by company name
      description: Returns matching ticker symbols for a given company name query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TickerSearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TickerSearchResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ScreenerRequest:
      type: object
      properties:
        tickers:
          type: array
          items:
            type: string
          description: List of ticker symbols to screen
        maxPE:
          type: number
          description: Maximum forward P/E ratio (default 30)
        maxPB:
          type: number
          description: Maximum price-to-book ratio (default 3)
        maxDebtToEquity:
          type: number
          description: Maximum debt-to-equity ratio as a percentage (default 100)
        minCurrentRatio:
          type: number
          description: Minimum current ratio (default 1.2)
        maxMarketCap:
          type: number
          description: Maximum market cap in millions USD (default 2000000)
      required:
        - tickers
    ScreenerResult:
      type: object
      properties:
        ticker:
          type: string
        companyName:
          type: string
        price:
          type: number
        dayChangePercent:
          type: number
          description: Regular market day change as a decimal (e.g. -0.023 = -2.3%)
        sixMonthChangePercent:
          type: number
          description: 6-month price change as a decimal (e.g. -0.275 = -27.5%)
        forwardPE:
          type: number
        priceToBook:
          type: number
        profitMargin:
          type: number
        debtToEquity:
          type: number
        currentRatio:
          type: number
        pfcfRatio:
          type: number
          description: Price-to-Free-Cash-Flow ratio (marketCap / freeCashflow TTM)
        freeCashflow:
          type: number
          description: Trailing twelve months free cash flow in USD
        nextEarningsDate:
          type: string
          description: ISO date string of next earnings date, or null if unavailable
        sector:
          type: string
        industry:
          type: string
        marketCap:
          type: number
        fiftyTwoWeekHigh:
          type: number
        fiftyTwoWeekLow:
          type: number
        altmanZScore:
          type: number
          description: "Altman Z-score: >2.99 safe zone, 1.81-2.99 grey zone, <1.81 distress zone"
        country:
          type: string
        state:
          type: string
        city:
          type: string
      required:
        - ticker
        - companyName
        - price
        - forwardPE
        - priceToBook
        - profitMargin
        - debtToEquity
        - currentRatio
    CommodityResult:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        price:
          type: number
        dayChangePercent:
          type: number
          description: Day change as a decimal (e.g. 0.002 = +0.2%)
        sixMonthChangePercent:
          type: number
          description: 6-month price change as a decimal
        prevClose:
          type: number
        isForwardContract:
          type: boolean
          description: True if this is the 3rd-month forward contract row
        baseGroup:
          type: string
          description: Front-month ticker this row belongs to (e.g. GC=F)
      required:
        - ticker
        - name
        - price
    ScreenerResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ScreenerResult"
        commodities:
          type: array
          items:
            $ref: "#/components/schemas/CommodityResult"
        screened:
          type: integer
          description: Total number of tickers screened
        passed:
          type: integer
          description: Number of tickers that passed the filter
        criteria:
          type: object
          properties:
            maxPE:
              type: number
            minMargin:
              type: number
          required:
            - maxPE
            - minMargin
        errors:
          type: array
          items:
            type: object
            properties:
              ticker:
                type: string
              error:
                type: string
            required:
              - ticker
              - error
      required:
        - results
        - commodities
        - screened
        - passed
        - criteria
        - errors
    QuoteRequest:
      type: object
      properties:
        tickers:
          type: array
          items:
            type: string
      required:
        - tickers
    StockQuote:
      type: object
      properties:
        ticker:
          type: string
        companyName:
          type: string
        price:
          type: number
        dayChangePercent:
          type: number
          description: Regular market day change as a decimal (e.g. -0.023 = -2.3%)
        sixMonthChangePercent:
          type: number
          description: 6-month price change as a decimal (e.g. -0.275 = -27.5%)
        forwardPE:
          type: number
        priceToBook:
          type: number
        profitMargin:
          type: number
        debtToEquity:
          type: number
        currentRatio:
          type: number
        pfcfRatio:
          type: number
          description: Price-to-Free-Cash-Flow ratio
        freeCashflow:
          type: number
          description: Trailing twelve months free cash flow in USD
        nextEarningsDate:
          type: string
          description: ISO date string of next earnings date, or null if unavailable
        sector:
          type: string
        industry:
          type: string
        marketCap:
          type: number
        fiftyTwoWeekHigh:
          type: number
        fiftyTwoWeekLow:
          type: number
        error:
          type: string
      required:
        - ticker
    QuoteResponse:
      type: object
      properties:
        quotes:
          type: array
          items:
            $ref: "#/components/schemas/StockQuote"
      required:
        - quotes
    TopStock:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        marketCap:
          type: number
        price:
          type: number
      required:
        - ticker
        - name
    TopMarketCapResponse:
      type: object
      properties:
        stocks:
          type: array
          items:
            $ref: "#/components/schemas/TopStock"
        total:
          type: integer
      required:
        - stocks
        - total
    TickerSearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Company name or partial ticker to search for
      required:
        - query
    TickerSearchResult:
      type: object
      properties:
        symbol:
          type: string
        shortname:
          type: string
        exchange:
          type: string
        quoteType:
          type: string
      required:
        - symbol
    TickerSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/TickerSearchResult"
      required:
        - results
    RateResult:
      type: object
      properties:
        name:
          type: string
          description: Human-readable instrument name (e.g. "10 Year UST")
        symbol:
          type: string
          description: Yahoo Finance ticker symbol
        quoteType:
          type: string
          enum: [yield, futures, etf]
          description: "yield: treasury yield index; futures: price of futures contract; etf: ETF price"
        value:
          type: number
          description: "Display value in natural units: yield in %, futures price, ETF price"
        displayValue:
          type: string
          description: Pre-formatted display string (e.g. '4.321%' or '$97.45')
        dayChange:
          type: number
          description: Absolute change in display units (yield pp, futures pts, ETF $)
        dayChangePercent:
          type: number
          description: Percentage change from previous close as a decimal
      required:
        - name
        - symbol
        - quoteType
        - value
        - displayValue
    BenchmarkResult:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        dayChange:
          type: number
        dayChangePercent:
          type: number
      required:
        - symbol
        - name
        - price
        - dayChange
        - dayChangePercent
    BenchmarksResponse:
      type: object
      properties:
        benchmarks:
          type: array
          items:
            $ref: "#/components/schemas/BenchmarkResult"
      required:
        - benchmarks
    RatesResponse:
      type: object
      properties:
        rates:
          type: array
          items:
            $ref: "#/components/schemas/RateResult"
      required:
        - rates