# lib/api-zod/src/generated/types/benchmarkResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BenchmarkResult {
  symbol: string;
  name: string;
  price: number;
  dayChange: number;
  dayChangePercent: number;
}