# lib/api-zod/src/generated/types/benchmarksResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BenchmarkResult } from "./benchmarkResult";

export interface BenchmarksResponse {
  benchmarks: BenchmarkResult[];
}